pyproject.toml
[project]
name = "endless8"
version = "0.1.0"
description = "pydantic-ai と claudecode-model を使用したコンテキスト効率の良いタスク実行ループエンジン"
readme = "README.md"
authors = [
    { name = "driller", email = "eleshis@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "claudecode-model",
    "duckdb>=1.4.3",
    "pydantic-ai>=1.46.0",
    "pyyaml>=6.0.0",
    "typer>=0.21.1",
]

[project.scripts]
e8 = "endless8.cli.main:app"

[build-system]
requires = ["uv_build>=0.9.24,<0.10.0"]
build-backend = "uv_build"

[tool.uv.sources]
claudecode-model = { git = "https://github.com/drillan/claudecode-model.git" }

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.14",
    "types-pyyaml>=6.0.12.20250915",
]

[tool.ruff]
target-version = "py313"
line-length = 88
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
]

[tool.ruff.lint.isort]
known-first-party = ["endless8"]

[tool.mypy]
python_version = "3.13"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_any_generics = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["duckdb.*", "claudecode_model.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-v --cov=endless8 --cov-report=term-missing"
